+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.9.1] - 2026-07-05
 
 ### Fixed
 
@@ -273,5 +273,6 @@ Please make sure to read the updated README for how to use the new version.
 ## 0.1.0 2009-03-28
 * Initial release.
 
+[2.9.1]: https://github.com/savonrb/nori/compare/v2.9.0...v2.9.1
 [2.9.0]: https://github.com/savonrb/nori/compare/v2.8.0...v2.9.0
 [2.8.0]: https://github.com/savonrb/nori/compare/v2.7.1...v2.8.0